feat: add VSCode settings for file associations and update graph printing loop to use int for index

+++ sources/structures/graph.hpp
@@ -45,7 +45,7 @@ public:
         for (int i = 0; i < numVertices; ++i) {
             std::cout << "Wierzcholek " << i << ": ";
             DynamicArray<Edge>* list = adjacencyList.get(i);
-            for (size_t j = 0; j < list->getSize(); ++j) {
+            for (int j = 0; j < list->getSize(); ++j) {
                 Edge e = list->get(j);
                 std::cout << "(" << e.to << ", " << e.weight << ") ";
             }
